README
I'm creating a 640 pixel wide and 480 pixel high sketch if im correct that creates white circles where ever your mouse is directed. Then when the mouse is clicked it creates a black circle until your mouse clicker is released. Thanks to the sample work provided I was able to create this.